Added base Hooks.java class to move before and after hooks
