+++ cukes/src/test/java/runner/CucumberRunnerTest.java
@@ -4,6 +4,6 @@ import io.cucumber.testng.AbstractTestNGCucumberTests;
 import io.cucumber.testng.CucumberOptions;
 
 @CucumberOptions(features = "src/test/resources/features", glue = "stepDefinitions",
-        monochrome = true, plugin = {"pretty", "html:target/cucumber"}, tags = "@outline")
+        monochrome = true, plugin = {"pretty", "html:target/cucumber"}, tags = "@regression")
 public class CucumberRunnerTest extends AbstractTestNGCucumberTests {
 }
