+++ cukes/src/test/java/stepDefinitions/Contact_Us_Test.java
@@ -1,36 +1,15 @@
 package stepDefinitions;
 
-import io.cucumber.java.After;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import io.cucumber.java.Before;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-//import org.openqa.selenium.chrome.ChromeOptions;
 import org.testng.Assert;
+import stepDefinitions.base.Hooks;
 
-public class Contact_Us_Test {
-
-    private WebDriver driver;
-    @Before("@contact-us")
-    public void setup() {
-        /*ChromeOptions options = new ChromeOptions();
-        options.addArguments("--no-sandbox");
-        options.addArguments("--disable-dev-shm-usage");
-        options.addArguments("--headless");
-        options.setBinary("/usr/local/share/chromedriver-linux64");
-        driver = new ChromeDriver(options);*/
-        driver = new ChromeDriver();
-        driver.manage().window().maximize();
-    }
-    @After("@contact-us")
-    public void tearDown() {
-        driver.quit();
-    }
+public class Contact_Us_Test extends Hooks {
 
     public String getRandomData(int length) {
         return RandomStringUtils.randomNumeric(length);
