+++ cukes/src/test/resources/features/Login.feature
@@ -1,4 +1,4 @@
-@login
+@login @regression
 Feature: Webdriver University - Login Feature
   This feature allows testing of login scenarios
 
@@ -6,16 +6,16 @@ Feature: Webdriver University - Login Feature
     Given I access the webdriver university login page
 
   Scenario: Validate Successful Login
-    When I enter a username "webdriver"
-    And I enter a password "webdriver123"
+    When I enter a username webdriver
+    And I enter a password webdriver123
     * I click on the login button
-    Then I should bee presented with the successful login message
+    Then I should be presented with the successful login message
 
   Scenario: Validate Unsuccessful Login
-    When I enter a username "webdriver"
-    And I enter a password "webdriver12"
+    When I enter a username webdriver
+    And I enter a password webdriver12
     * I click on the login button
-    Then I should bee presented with the unsuccessful login message
+    Then I should be presented with the unsuccessful login message
 
   @outline
   Scenario Outline: Validate Successful & Unsuccessful login message
